♻️ Move DefaultStepOptions to opts
