+++ step_options.go
@@ -1,8 +1,6 @@
 package sqlslog
 
 import (
-	"log/slog"
-
 	"github.com/akm/sql-slog/opts"
 )
 
@@ -14,27 +12,9 @@ type (
 )
 
 var (
-	NewStepOptions = opts.NewStepOptions
+	NewStepOptions     = opts.NewStepOptions
+	DefaultStepOptions = opts.DefaultStepOptions
 
 	StepLogMsgWithEventName    = opts.StepLogMsgWithEventName
 	StepLogMsgWithoutEventName = opts.StepLogMsgWithoutEventName
 )
-
-func DefaultStepOptions(formatter StepLogMsgFormatter, name string, completeLevel Level, errHandlers ...func(error) (bool, []slog.Attr)) *StepOptions {
-	var startLevel Level
-	switch completeLevel { // nolint:exhaustive
-	case LevelError:
-		startLevel = LevelInfo
-	case LevelInfo:
-		startLevel = LevelDebug
-	case LevelDebug:
-		startLevel = LevelTrace
-	default:
-		startLevel = LevelVerbose
-	}
-	r := NewStepOptions(formatter, name, startLevel, LevelError, completeLevel)
-	if len(errHandlers) > 0 {
-		r.ErrorHandler = errHandlers[0]
-	}
-	return r
-}
